+++ CHANGELOG.md
@@ -7,13 +7,26 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Minor Bug Fixes
 
 ### Oneview Features supported
+- Enclosure
+- Enclosure Group
+- Ethernet Network
+- FC Network
+- FCOE Network
+- Interconnects
+- Interconnect Types
+- Logical Enclosure
 - Logical Interconnect
+- Network Set
 - Scope
 - Server Profile
 - Server Profile Template
+- Tasks
 
 #### Bug fixes & Enhancements:
-- [#239] (https://github.com/HewlettPackard/oneview-golang/issues/239) CreateProfileFromTemplate not working w OV 5.20
+- [#239] (https://github.com/HewlettPackard/oneview-golang/issues/239) CreateProfileFromTemplate not working with OV 5.20
+
+### Major changes:
+- Refactored SDK to take default API version from Oneview appliance.
 
 ### Major changes:
 - Refactored SDK to take default API version from Oneview appliance.