+++ CHANGELOG.md
@@ -16,17 +16,18 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Interconnect Types
 - Logical Enclosure
 - Logical Interconnect
+- Logical Interconnect Group
 - Network Set
 - Scope
+- Server Certificates
+- Server Hardware
+- Server Hardware Types
 - Server Profile
 - Server Profile Template
 - Tasks
 
 #### Bug fixes & Enhancements:
 - [#239] (https://github.com/HewlettPackard/oneview-golang/issues/239) CreateProfileFromTemplate not working with OV 5.20
-
-### Major changes:
-- Refactored SDK to take default API version from Oneview appliance.
 
 ### Major changes:
 - Refactored SDK to take default API version from Oneview appliance.